+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+###
+- file description field
+
 ### Changed
 - keep time information in loaded tracks/routes/waypoints, written in saved files